refactor: Shard the crypto files to prevent circular dep warnings
