+++ src/crypto/crypto.ts
@@ -15,15 +15,9 @@
 
 import {BrowserCrypto} from './browser/crypto';
 import {NodeCrypto} from './node/crypto';
+import {Crypto} from './shared';
 
-export interface JwkCertificate {
-  kty: string;
-  alg: string;
-  use?: string;
-  kid: string;
-  n: string;
-  e: string;
-}
+export {Crypto, JwkCertificate} from './shared';
 
 export interface CryptoSigner {
   update(data: string): void;
@@ -37,41 +31,6 @@ export interface CryptoSigner {
 // SubtleCrypto methods return promises, we must make those
 // methods return promises here as well, even though in Node.js
 // they are synchronous.
-export interface Crypto {
-  sha256DigestBase64(str: string): Promise<string>;
-  randomBytesBase64(n: number): string;
-  verify(
-    pubkey: string | JwkCertificate,
-    data: string | Buffer,
-    signature: string
-  ): Promise<boolean>;
-  sign(
-    privateKey: string | JwkCertificate,
-    data: string | Buffer
-  ): Promise<string>;
-  decodeBase64StringUtf8(base64: string): string;
-  encodeBase64StringUtf8(text: string): string;
-  /**
-   * Computes the SHA-256 hash of the provided string.
-   * @param str The plain text string to hash.
-   * @return A promise that resolves with the SHA-256 hash of the provided
-   *   string in hexadecimal encoding.
-   */
-  sha256DigestHex(str: string): Promise<string>;
-
-  /**
-   * Computes the HMAC hash of a message using the provided crypto key and the
-   * SHA-256 algorithm.
-   * @param key The secret crypto key in utf-8 or ArrayBuffer format.
-   * @param msg The plain text message.
-   * @return A promise that resolves with the HMAC-SHA256 hash in ArrayBuffer
-   *   format.
-   */
-  signWithHmacSha256(
-    key: string | ArrayBuffer,
-    msg: string
-  ): Promise<ArrayBuffer>;
-}
 
 export function createCrypto(): Crypto {
   if (hasBrowserCrypto()) {
@@ -86,20 +45,4 @@ export function hasBrowserCrypto() {
     typeof window.crypto !== 'undefined' &&
     typeof window.crypto.subtle !== 'undefined'
   );
-}
-
-/**
- * Converts an ArrayBuffer to a hexadecimal string.
- * @param arrayBuffer The ArrayBuffer to convert to hexadecimal string.
- * @return The hexadecimal encoding of the ArrayBuffer.
- */
-export function fromArrayBufferToHex(arrayBuffer: ArrayBuffer): string {
-  // Convert buffer to byte array.
-  const byteArray = Array.from(new Uint8Array(arrayBuffer));
-  // Convert bytes to hex string.
-  return byteArray
-    .map(byte => {
-      return byte.toString(16).padStart(2, '0');
-    })
-    .join('');
 }
